requirements.txt
matplotlib
numpy
palettable
pandas
opencv-python
tqdm
numba
pyyaml
swifter
seaborn
scicomap
pyarrow
fastparquet
sklearn
tensorflow